refactor: apis 폴더 이름 변경
- APIs => apiRequest
